### 2025-07-08 #8554 @editorialbot
[PRE REVIEW]: Ivy v1: A Data Acquisition System for Distributed Sensor Networks Supporting Remote Configuration and Software Updates
**Submitting author:** @dostuffthatmatters (<a href="http://orcid.org/0000-0002-2948-2993">Moritz Makowski</a>)
**Repository:** https://github.com/tum-esm/ivy
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
