### 2025-07-08 1d690ce
Update paper
Shorten paper from 2047 words to 1051 words.
