+++ paper/paper.md
@@ -39,75 +39,51 @@ bibliography: paper.bib
 
 # Summary
 
-Distributed sensor networks monitor environmental conditions at remote locations. Each sensor node produces various time series data streams and system logs and sends this data to a central backend. A "sensor node" is an autonomous system collecting data from one or multiple sensors connected to it, operated 24/7 by a Data Acquisition System (DAS).
+Distributed sensor networks monitor environmental conditions at remote locations. Each sensor node produces various time series data streams and system logs and sends this data to a central backend. A "sensor node" is an autonomous system collecting data from one or multiple sensors connected to it, operated 24/7 by a Data Acquisition System (DAS). As part of the ICOS Cities PAUL Project [@paul], our group has developed a network of 20 autonomous GHG (greenhouse gas) sensor nodes [@midcost].
 
-Starting in 2022, our group has developed and deployed a network of 20 autonomously operating GHG (greenhouse gas) sensor nodes as part of the ICOS Cities PAUL Project (Pilot Applications In Urban Landscapes) [@paul;@midcost]. Visiting each sensor site on every configuration change or software update is not feasible when operating 20 or more sensor nodes. To overcome this, we developed a DAS that supports remote configuration and software updates [@hermes]. The sensor node of Hermes is a Python-based DAS that runs on a Raspberry Pi and operates various sensors and actuators 24/7. The ability to update the DAS remotely enabled frequent improvements in sensor software after deployment.
+Since the software architecture of a DAS is independent of a specific sensor network, we present Ivy – a boilerplate for a DAS that supports configuring and updating itself remotely. Research teams building a new sensor network can use Ivy as a base for their own network-specific DAS. Ivy provides the core architecture, and users of the boilerplate only have to plug in their network-specific hardware logic to make it operational. Ivy is designed to be owned and customized: one can switch to another documentation system, add another backend, or remove unused parts of the boilerplate.
 
-Since the software architecture of a DAS is independent of a specific sensor network, we present Ivy – a boilerplate for a DAS that supports configuring and updating itself remotely. Research teams building a new sensor network can use Ivy as a base for their own network-specific DAS. Ivy provides the core architecture and project organization – users of the boilerplate only have to plug in their network-specific hardware logic to make it operational, instead of building a DAS from scratch. Ivy is designed to be owned and customized: one can switch to another documentation system, add another backend, or remove unused parts of the boilerplate.
+\vspace{4mm}
 
-![The main components of a sensor network. Each sensor node is operated by a DAS like Ivy.\label{fig:architecture1}](figures/ivy-architecture-fig-1.png){ width=94.3% }
+![System architecture of a distributed sensor network based on a DAS like Ivy.\label{fig:architecture1}](figures/ivy-architecture-fig-1.png){ width=94.3% }
 
-\autoref{fig:architecture1} shows the system architecture of a sensor network using an autonomous DAS like Ivy. This architecture template is the result of many iterations of the sensor networks our research group has built and operated [@Dietrich2021;@hermes;@Aigner2023]. This paper presents Ivy version 1, which has been built to the best of our knowledge about how a reliable and extendable DAS architecture can look. We aim to continuously improve and extend the Ivy boilerplate based on the feedback from the community resulting from this publication.
+The architecture of Ivy shown in \autoref{fig:architecture1} results from many iterations of the sensor networks our research group has built and operated [@Dietrich2021;@hermes;@Aigner2023]. This publication aims to share a reference architecture of how a reliable DAS can be built, not claiming that Ivy is the only architecture for this use case.
 
[…]
