+++ paper/paper.bib
@@ -463,3 +463,13 @@ year = {2025}
   month = feb,
   pages = {14}
 }
+
+@article{Wenzel2025,
+  title = {Towards high-resolution air pollutants sensing through dense low-cost sensor networks &amp;#8211; a case study in Munich},
+  url = {http://dx.doi.org/10.5194/egusphere-egu25-16784},
+  DOI = {10.5194/egusphere-egu25-16784},
+  publisher = {Copernicus GmbH},
+  author = {Wenzel,  Adrian and Chen,  Jia and Klama,  Tobias and B\"{o}hm,  Felix and Angleitner,  Moritz and Lobmaier,  Reinhard},
+  year = {2025},
+  month = mar 
+}
