### 2025-08-13 #8554 @danielskatz
👋 @dostuffthatmatters - We have finished the scope review, which your submission passed. We also changed its track, so I will be looking for an editor for it next.
