### 2025-08-24 #8554 @ChitlangeSahas
> Greetings 👋 [@cppraveen](https://github.com/cppraveen), [@brpat07](https://github.com/brpat07), [@r3w0p](https://github.com/r3w0p), [@ChitlangeSahas](https://github.com/ChitlangeSahas)
>
> would you be interested in reviewing this submission for JOSS?

@nkrusch  - thank you for reaching out, yes would love to! I'll begin by brushing up on the reviewer guidelines and then begin reading this paper, will submit the review as a comment here 👍
