### 2025-08-24 #8554 @cppraveen
> Greetings 👋 [@cppraveen](https://github.com/cppraveen), [@brpat07](https://github.com/brpat07), [@r3w0p](https://github.com/r3w0p), [@ChitlangeSahas](https://github.com/ChitlangeSahas)
>
> would you be interested in reviewing this submission for JOSS?

@nkrusch - Sure, i'm excited to contribute
