### 2025-08-30 tum-esm/ivy#5 @brpat07
[Improvement][Doc]: Provide Accessible Installation and Usage Guide
It would be very helpful for users to have clear and accessible installation instructions, ideally included in the repo README or a main documentation page.

Currently, there are no dedicated instructions on how to install or set up the project—aside from hints in main.py or in [src.md](https://github.com/tum-esm/ivy/blob/main/docs/pages/api-reference/src.md?utm_source=chatgpt.com), which are harder to find.

Also, some specific gaps in the existing guidance:
1. There’s no mention of requiring pdm (or alternatively pip) to install dependencies.
2. The documentation does not explain how the template config file can be used to create or load a configuration.

#### Suggested improvements
It would be great to have a dedicated, easily discoverable markdown doc (linked from the README) that includes:
- Installation instructions (dependencies, setup steps, or the script to run if any)
- Instructions to run / integrate
- Instructions to test (this may already exist, but could be linked or clarified if it does)

(https://github.com/openjournals/joss-reviews/issues/8862)
