### 2025-08-30 #8862 @brpat07
@dostuffthatmatters created this issue (https://github.com/tum-esm/ivy/issues/5) for adding installation and other instructions
