### 2025-08-30 tum-esm/ivy#5 @dostuffthatmatters
Hi @brpat07

Thank you for the suggestion!

You are right, the actual steps to set up a new system are implicit and distributed over the documentation. I will add a new page describing the full setup of a new system!

I will ping you here soon.
