### 2025-09-02 tum-esm/ivy#5 @dostuffthatmatters
Hi @brpat07

I have added a new page "Getting Started" (https://tum-esm-ivy.netlify.app/getting-started) explaining how to 1. set up Ivy on a development machine and 2. how to set up a new production sensor node running an Ivy-based DAS.

Let me know if you think more information can be added.

Best,
Moritz Makowski

@ChitlangeSahas @cppraveen FYI
