### 2025-09-02 tum-esm/ivy#5 @ChitlangeSahas
Thank you @dostuffthatmatters , I finished reading your paper last night and will follow the getting started steps today to run through the actual software installation and runtime procedures.
