### 2025-09-02 d3752a0
Update paper
* Remove duplicate citations
* Use group name for repositories of equal contribution of all authors
