+++ paper/paper.bib
@@ -210,8 +210,8 @@
 }
 
 @software{hermes,
-  author = {Aigner, Patrick and Böhm, Felix and Frölich, Lars and Makowski, Moritz and Schmitt, Adrian},
-  title = {Hermes},
+  author = {{TUM ESM} and Aigner, Patrick and Böhm, Felix and Frölich, Lars and Makowski, Moritz and Schmitt, Adrian and Chen, Jia},
+  title = {Hermes, Professorship of Environmental Sensing and Modeling, Technical University of Munich},
   url = {https://github.com/tum-esm/hermes},
   version = {1.0.1},
   date = {2024-10-14}
@@ -281,8 +281,8 @@
 }
 
 @software{acropolis-edge,
-  author = {Aigner, Patrick and Frölich, Lars},
-  title = {{ACROPOLIS-edge}},
+  author = {{TUM ESM} and Aigner, Patrick and Böhm, Felix and Frölich, Lars and Makowski, Moritz and Schmitt, Adrian and Chen, Jia}, 
+  title = {ACROPOLIS-edge, Professorship of Environmental Sensing and Modeling, Technical University of Munich},
   url = {https://github.com/tum-esm/ACROPOLIS-edge},
   version = {0.0.95},
   date = {2025-04-04}
