### 2025-09-03 #8862 @ChitlangeSahas
Added some minor comments @dostuffthatmatters over in https://github.com/tum-esm/ivy/issues/5#issuecomment-3247535458.

I'd be happy to make the PR with the suggestions, if you would like that, let me know!
