### 2025-09-04 tum-esm/ivy#5 @brpat07
@dostuffthatmatters I encountered the same issue as @ChitlangeSahas mentioned in his comment.

My recommendation would be to add a small starter-script (either in python or bash) which would take care of basic setup (like backend setup, etc).

For the env (IVY_ROOT_DIR) setup, either you can include in the starter-script
or better way to do is
to update the code to identify the parent dir instead of hardcoding in `~/Documents/{NAME}` (here https://github.com/tum-esm/ivy/blob/main/src/constants.py#L24)
