### 2025-09-04 c283bfe
Improve setup scripts (1)
Move test scripts from pdm scripts to bash files
