+++ pyproject.toml
@@ -67,19 +67,6 @@ markers = [
 ]
 
 [tool.pdm.scripts]
-quick-tests = { shell = """
-  export TS="$(date +%s)" && export IVY_ROOT_DIR="/tmp/ci-tests-ivy-root-dir-$TS" && mkdir $IVY_ROOT_DIR &&
-  pytest --verbose -m "quick" --cov=$PWD/src --exitfirst tests/ && rm -rf /tmp/ci-tests-ivy* &&
-  pkill -f "Python"
-  """ }
-ci-tests = { shell = """
-  export TS="$(date +%s)" && export IVY_ROOT_DIR="/tmp/ci-tests-ivy-root-dir-$TS" && mkdir $IVY_ROOT_DIR &&
-  pytest --verbose -m "quick or updater" --cov=$PWD/src --exitfirst tests/ && rm -rf /tmp/ci-tests-ivy* &&
-  pkill -f "Python"
-  """ }
-integration-tests = { shell = """
-  pytest --verbose -m "integration" --cov=$PWD/src --exitfirst tests/
-  """ }
 start-mosquitto = { shell = """
   mosquitto_passwd -c -b $PWD/tests/updater/data/mosquitto.passwords test_username test_password &&
   echo "listener 1883\\nallow_anonymous false\\npassword_file $PWD/tests/updater/data/mosquitto.passwords" > $PWD/tests/updater/data/mosquitto.conf &&
