### 2025-09-04 9cb6598
Improve setup scripts (2)
Move mosquitto scripts from pdm scripts to bash files
