+++ scripts/mosquitto_stop.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+# stops the mosquitto broker
+pkill -f "mosquitto -d"
