### 2025-09-04 e97a1f6
Improve setup scripts (3)
Move thingsboard scripts from pdm scripts to bash files
