### 2025-09-04 fc8f4dc
Improve setup scripts (4)
Add script to setup a production node
