+++ scripts/setup_production.py
@@ -1,0 +1,183 @@
+#!/bin/python3
+
+import contextlib
+import json
+import os
+import re
+import shutil
+import sys
+import tarfile
+from typing import Generator
+import urllib.request
+
+BOLD_CODE = "\033[1m"
+GREEN_CODE = "\033[32m"
+YELLOW_CODE = "\033[33m"
+RED_CODE = "\033[31m"
+GRAY_CODE = "\033[90m"
+RESET_CODE = "\033[0m"
+
+
+@contextlib.contextmanager
+def section(label: str) -> Generator[None, None, None]:
+    print(f"{BOLD_CODE}{label}: Starting{RESET_CODE}")
+    print(GRAY_CODE)
+    yield
+    print(RESET_CODE)
+    print(f"{GREEN_CODE}{label}: Finished{RESET_CODE}")
+
+
+def error(message: str) -> None:
+    print(f"{RED_CODE}ERROR: {message}{RESET_CODE}")
+    sys.exit(1)
+
+
+def get_required_env_var(var_name: str) -> str:
+    """Get required environment variable or exit with error."""
+    value = os.getenv(var_name)
+    if value is None:
+        error(f"Required environment variable {var_name} is not set")
+    return value.strip()
+
+
+def load_env_vars() -> dict[str, str]:
+    with section("Loading environment variables"):
+        env_vars = {
+            "project_name": get_required_env_var("IVY_PROJECT_NAME"),
+            "git_repository": get_required_env_var("IVY_GIT_REPOSITORY"),
+            "git_provider": get_required_env_var("IVY_GIT_PROVIDER"),
+            "version": get_required_env_var("IVY_VERSION"),
+            "system_identifier": get_required_env_var("IVY_SYSTEM_IDENTIFIER"),
+            "config_filepath": get_required_env_var("IVY_CONFIG_FILEPATH"),
+            "package_manager": get_required_env_var("IVY_PACKAGE_MANAGER"),
+        }
+
+        if not env_vars["project_name"]:
+            error("Project name cannot be empty")
+
+        if not env_vars["git_repository"].startswith("https://"):
+            error("Git repository must start with 'https://'")
+
+        if not env_vars["git_provider"] in ["github", "gitlab"]:
+            error("Git repository must be from either 'github' or 'gitlab'")
+
+        if not env_vars["version"].startswith("v"):
+            env_vars["version"] = f"v{env_vars['version']}"
+
+        if not re.match(r"^[a-z0-9-]+$", env_vars["system_identifier"]):
+            error("System identifier can only contain lowercase letters, numbers, and hyphens")
+
+        if not os.path.exists(env_vars["config_filepath"]):
+            error(f"Config source file not found: {env_vars['config_filepath']}")
+
+        if not os.path.isfile(env_vars["config_filepath"]):
+            error(f"Config source must be a file, not a directory: {env_vars['config_filepath']}")
+
+        if env_vars["package_manager"] not in ["pip", "pdm", "uv"]:
+            error("Package manager must be one of: pip, pdm, uv")
+
+        if sys.version_info.major != 3 or sys.version_info.minor < 10:
+            error("Python version must be 3.10 or higher")
+
+        print(json.dumps(env_vars, indent=4))
+
+    return env_vars
+
+
+def setup_directories(config: dict[str, str]) -> tuple[str, str, str]:
+    with section("Setting up directories"):
+        documents_dir = os.path.join(os.path.expanduser("~"), "Documents")
+        project_dir = os.path.join(documents_dir, config["project_name"])
+        version_dir = os.path.join(project_dir, config["version"].lstrip("v"))
+        for path in [documents_dir, project_dir, version_dir]:
+            if os.path.isfile(path):
[…]
