### 2025-09-04 7fd8f73
Improve setup scripts (5)
Use new production setup script in getting started guide
