### 2025-09-04 8855c86
Improve setup scripts (6)
Add script to setup a development computer
