+++ scripts/setup_development.py
@@ -1,0 +1,245 @@
+#!/bin/python3
+
+from typing import Callable, Generator, Optional
+import contextlib
+import json
+import os
+import re
+import socket
+import sys
+
+BOLD_CODE = "\033[1m"
+GREEN_CODE = "\033[32m"
+YELLOW_CODE = "\033[33m"
+RED_CODE = "\033[31m"
+GRAY_CODE = "\033[90m"
+RESET_CODE = "\033[0m"
+
+
+@contextlib.contextmanager
+def section(label: str) -> Generator[Callable[[str], None], None, None]:
+    print(f"{BOLD_CODE}{label}: Starting{RESET_CODE}")
+    print(GRAY_CODE)
+    yield lambda message: print(f"{RESET_CODE}{message}{GRAY_CODE}")
+    print(RESET_CODE)
+    print(f"{GREEN_CODE}{label}: Finished{RESET_CODE}")
+
+
+def error(message: str) -> None:
+    print(f"{RED_CODE}ERROR: {message}{RESET_CODE}")
+    sys.exit(1)
+
+
+def format_warning(message: str) -> str:
+    return f"{YELLOW_CODE}WARNING: {message}{RESET_CODE}"
+
+
+def get_validated_input(
+    prompt: str,
+    conditions: list[tuple[Callable[[str], bool], str]],
+) -> str:
+    while True:
+        output = input(f"{BOLD_CODE}{prompt}\n>{RESET_CODE} ").strip("\t\n ")
+        valid_output: bool = True
+        for condition, message in conditions:
+            if not condition(output):
+                print(f"{RED_CODE}Invalid input: {message}{RESET_CODE}")
+                valid_output = False
+                break
+        if valid_output:
+            return output
+
+
+# GET USER INPUTS
+
+PROJECT_NAME = get_validated_input(
+    "What is the name of your DAS? (instead of `ivy`)",
+    conditions=[
+        (lambda s: len(s) > 0, "Project name cannot be empty"),
+    ],
+)
+
+GIT_REPOSITORY = get_validated_input(
+    "Where is your git repository hosted? (instead of `https://github.com/tum-esm/ivy`)",
+    conditions=[
+        (
+            lambda s: s.startswith("https://") or s.startswith("git@"),
+            "Git repository must start with 'https://' or 'git@'",
+        ),
+    ],
+)
+
+PACKAGE_MANAGER = get_validated_input(
+    "How do you want to install your dependencies? (pip | pdm | uv)",
+    conditions=[
+        (lambda s: s in ["pip", "pdm", "uv"], "Must be one of 'pip', 'pdm', or 'uv'"),
+    ],
+)
+
+HOSTNAME = socket.gethostname().replace("_", "-").replace(" ", "-").lower()
+SYSTEM_IDENTIFIER = get_validated_input(
+    f'How do you want to name this computer? (if empty, uses "{HOSTNAME}")',
+    conditions=[
+        (lambda s: len(s) <= 512, "System identifier cannot be empty"),
+        (
+            lambda s: re.match(r"^[a-z0-9-]+$", s),
+            "Only lowercase letters, numbers, and hyphens are allowed",
+        ),
+    ],
+)
+if SYSTEM_IDENTIFIER == "":
+    SYSTEM_IDENTIFIER = HOSTNAME
+
+CONFIGURE_BACKEND = (
+    get_validated_input(
+        "Do you want to configure an MQTT broker? (yes | no)",
+        conditions=[
+            (
+                lambda s: s.lower() in ["yes", "no", "y", "n"],
+                "Must be 'yes'/'y' or 'no'/'n'",
+            ),
+        ],
+    )
+    .lower()
[…]
