### 2025-09-04 efd6b0e
Improve setup scripts (7)
Mention new development setup script in getting started guide
