### 2025-09-04 2fbbeb8
Improve setup scripts (8)
Add option to feed env vars into dev setup script
