### 2025-09-04 73311a5
Improve setup scripts (9)
Test setup scripts with static type tests
