+++ tests/quick/test_static_types.py
@@ -14,7 +14,8 @@ def test_static_types() -> None:
     _rm(".mypy_cache/3.*/src")
     _rm(".mypy_cache/3.*/tests")
     _rm(".mypy_cache/3.*/run.*")
+    _rm(".mypy_cache/3.*/scripts")
     _rm(".mypy_cache/3.*/docs/scripts")
 
-    for path in ["src/", "tests/", "run.py", "docs/scripts/"]:
+    for path in ["src/", "tests/", "run.py", "scripts", "docs/scripts/"]:
         assert os.system(f"cd {src.constants.PROJECT_DIR} && {sys.executable} -m mypy {path}") == 0
