### 2025-09-04 3eaef48
Improve setup scripts (11)
Rename codebase test ci script
Add development test script
