+++ .github/workflows/test-codebase.yaml
@@ -1,4 +1,4 @@
-name: "test"
+name: "test-codebase"
 on:
   push:
     branches:
