+++ .github/workflows/test-development-setup.yaml
@@ -1,0 +1,57 @@
+name: "test-development-setup"
+on:
+  push:
+    branches:
+      - main
+  pull_request:
+    branches:
+      - main
+
+jobs:
+  test:
+    runs-on: ubuntu-24.04
+    strategy:
+      matrix:
+        python_version: ["3.10"]
+    defaults:
+      run:
+        shell: bash
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v3
+      - name: Set up Python ${{ matrix.python_version }}
+        uses: actions/setup-python@v5
+        with:
+          python-version: ${{ matrix.python_version }}
+          cache: "pip"
+
+      - name: Setup PDM
+        uses: pdm-project/setup-pdm@v4
+
+      - name: Run development setup script
+        env:
+          IVY_PROJECT_NAME: ${{ github.event.repository.name }}
+          IVY_GIT_REPOSITORY: https://github.com/${{ github.repository }}
+          IVY_PACKAGE_MANAGER: pdm
+          IVY_SYSTEM_IDENTIFIER: github-actions-node
+          IVY_CONFIGURE_BACKEND: yes
+          IVY_PROCEED: yes
+        run: curl -s https://raw.githubusercontent.com/${{ github.repository }}/refs/heads/${{ github.ref_name }}/scripts/setup_development.py | python3.12
+
+      - name: Install mosquitto dependencies
+        run: |
+          sudo apt-add-repository ppa:mosquitto-dev/mosquitto-ppa -y
+          sudo apt-get update -y
+          sudo apt-get install -y mosquitto-clients mosquitto
+
+      - name: Start mosquitto broker
+        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        run: ./scripts/mosquitto_start.sh
+
+      - name: Test whether broker is running
+        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        run: sleep 5 && ./scripts/mosquitto_test.sh
+
+      - name: Run integration tests
+        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        run: ./scripts/run_tests_integration.sh
