### 2025-09-04 e0be7a4
Improve setup scripts (12)
Exit bash scripts on first error
Add script to test mosquitto broker
