+++ scripts/run_thingsboard.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# fail on first error
+set -e errexit
+
 # parent directory of the `README.md` file
 export PROJECT_DIR=$(dirname $(dirname $(realpath $0)))
 
