+++ scripts/mosquitto_stop.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 
+# fail on first error
+set -e errexit
+
 # stops the mosquitto broker
 pkill -f "mosquitto -d"
