### 2025-09-04 16bc233
Improve setup scripts (13)
Debug CI scripts
Fix mypy issues
Disable codebase tests for a bit because it works
