+++ .github/workflows/test-development-setup.yaml
@@ -32,11 +32,12 @@ jobs:
         env:
           IVY_PROJECT_NAME: ${{ github.event.repository.name }}
           IVY_GIT_REPOSITORY: https://github.com/${{ github.repository }}
+          IVY_BRANCH_NAME: ${{ github.head_ref }}
           IVY_PACKAGE_MANAGER: pdm
-          IVY_SYSTEM_IDENTIFIER: github-actions-node
+          IVY_SYSTEM_IDENTIFIER: github-actions-node-development
           IVY_CONFIGURE_BACKEND: yes
           IVY_PROCEED: yes
-        run: curl -s https://raw.githubusercontent.com/${{ github.repository }}/refs/heads/${{ github.ref_name }}/scripts/setup_development.py | python3.12
+        run: curl -s https://raw.githubusercontent.com/${{ github.repository }}/refs/heads/${{ github.head_ref }}/scripts/setup_development.py | python3.10
 
       - name: Install mosquitto dependencies
         run: |
@@ -45,13 +46,13 @@ jobs:
           sudo apt-get install -y mosquitto-clients mosquitto
 
       - name: Start mosquitto broker
-        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        working-directory: /home/runner/Documents/${{ github.event.repository.name }}/dev
         run: ./scripts/mosquitto_start.sh
 
       - name: Test whether broker is running
-        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        working-directory: /home/runner/Documents/${{ github.event.repository.name }}/dev
         run: sleep 5 && ./scripts/mosquitto_test.sh
 
       - name: Run integration tests
-        working-directory: ~/Documents/${{ github.event.repository.name }}/dev
+        working-directory: /home/runner/Documents/${{ github.event.repository.name }}/dev
         run: ./scripts/run_tests_integration.sh
