### 2025-09-04 bacdff5
Improve setup scripts (14)
Refine production setup script for CI tests
