+++ scripts/setup_production.py
@@ -1,13 +1,12 @@
 #!/bin/python3
 
-from typing import Generator
+from typing import Generator, Optional
 import contextlib
 import json
 import os
 import re
 import shutil
 import sys
-import tarfile
 import urllib.request
 
 BOLD_CODE = "\033[1m"
@@ -41,7 +40,7 @@ def get_required_env_var(var_name: str) -> str:
     return value.strip()
 
 
-def load_env_vars() -> dict[str, str]:
+def load_config() -> dict[str, str]:
     with section("Loading environment variables"):
         env_vars = {
             "project_name": get_required_env_var("IVY_PROJECT_NAME"),
@@ -62,8 +61,8 @@ def load_env_vars() -> dict[str, str]:
         if not env_vars["git_provider"] in ["github", "gitlab"]:
             error("Git repository must be from either 'github' or 'gitlab'")
 
-        if not env_vars["version"].startswith("v"):
-            env_vars["version"] = f"v{env_vars['version']}"
+        if not (re.match(r"^\d+\.") or env_vars["version"].startswith("/")):
+            error("Version must either start with a number (e.g. '1.0.0') or be a an absolute path")
 
         if not re.match(r"^[a-z0-9-]+$", env_vars["system_identifier"]):
             error("System identifier can only contain lowercase letters, numbers, and hyphens")
@@ -89,7 +88,10 @@ def setup_directories(config: dict[str, str]) -> tuple[str, str, str]:
     with section("Setting up directories"):
         documents_dir = os.path.join(os.path.expanduser("~"), "Documents")
         project_dir = os.path.join(documents_dir, config["project_name"])
-        version_dir = os.path.join(project_dir, config["version"].lstrip("v"))
+        version_dir: str = "/tmp/thisdoesdefinitelyhopefullynotexist"
+        if not config["version"].startswith("v/"):
+            version_dir = os.path.join(project_dir, config["version"].lstrip("v"))
+
         for path in [documents_dir, project_dir, version_dir]:
             if os.path.isfile(path):
                 error(f"{path} is a file, not a directory. Please remove it.")
@@ -101,7 +103,18 @@ def setup_directories(config: dict[str, str]) -> tuple[str, str, str]:
         return documents_dir, project_dir, version_dir
 
 
-def download_and_codebase(config: dict[str, str], version_dir: str) -> None:
+def download_codebase(config: dict[str, str], project_dir: str) -> Optional[str]:
+    if config["version"].startswith("/"):
+        print(
+            f"Skipping download of codebase when path of latest codebase is given ({config['version']})"
+        )
+        assert os.path.isdir(config["version"])
+        with open(os.path.join(config["version"], "config", "config.template.json"), "r") as f:
+            version = json.load(f)["general"]["software_version"]
+        version_dir = os.path.join(project_dir, version)
+        os.rename(config["version"], version_dir)
+        return version
+
     with section(f"Downloading version {config['version']}"):
         git_repo = config["git_repository"]
         if git_repo.endswith(".git"):
@@ -126,10 +139,10 @@ def download_and_codebase(config: dict[str, str], version_dir: str) -> None:
             urllib.request.urlretrieve(download_url, archive_path)
         except Exception as e:
             error(f"Failed to download {download_url}: {e}")
-        with tarfile.open(archive_path, "r:gz") as tar:
-            tar.extractall("/tmp")
+        os.system(f"cd /tmp && tar -xzf {archive_path.split('/')[-1]}")
         os.rename(f"/tmp/{extracted_name}", version_dir)
         os.remove(archive_path)
+        return None
 
 
[…]
