### 2025-09-04 4fcdc33
Improve setup scripts (15)
Add CI job to test production setup
