### 2025-09-04 a4c1109
Improve setup scripts (16)
Debug CI script
