### 2025-09-04 72e6320
Improve setup scripts (17)
Refactor production setup script to not use functions
