+++ scripts/setup_production.py
@@ -1,6 +1,6 @@
 #!/bin/python3
 
-from typing import Generator, Optional
+from typing import Callable, Generator
 import contextlib
 import json
 import os
@@ -18,9 +18,9 @@ RESET_CODE = "\033[0m"
 
 
 @contextlib.contextmanager
-def section(label: str) -> Generator[None, None, None]:
+def section(label: str) -> Generator[Callable[[str], None], None, None]:
     print(f"{BOLD_CODE}{label}: Starting{RESET_CODE}{GRAY_CODE}")
-    yield
+    yield lambda message: print(f"{RESET_CODE}{message}{GRAY_CODE}")
     print(f"{RESET_CODE}{GREEN_CODE}{label}: Finished{RESET_CODE}")
 
 
@@ -38,9 +38,11 @@ def get_required_env_var(var_name: str) -> str:
     return value.strip()
 
 
-def load_config() -> dict[str, str]:
-    with section("Loading environment variables"):
-        env_vars = {
+if __name__ == "__main__":
+    print("🚀 Starting Ivy Production Setup")
+
+    with section("Loading environment variables") as localprint:
+        config = {
             "project_name": get_required_env_var("IVY_PROJECT_NAME"),
             "git_repository": get_required_env_var("IVY_GIT_REPOSITORY"),
             "git_provider": get_required_env_var("IVY_GIT_PROVIDER"),
@@ -50,49 +52,47 @@ def load_config() -> dict[str, str]:
             "package_manager": get_required_env_var("IVY_PACKAGE_MANAGER"),
         }
 
-        if not env_vars["project_name"]:
+        if not config["project_name"]:
             error("Project name cannot be empty")
 
-        if not env_vars["git_repository"].startswith("https://"):
+        if not config["git_repository"].startswith("https://"):
             error("Git repository must start with 'https://'")
 
-        if not env_vars["git_provider"] in ["github", "gitlab"]:
+        if not config["git_provider"] in ["github", "gitlab"]:
             error("Git repository must be from either 'github' or 'gitlab'")
 
-        if not (re.match(r"^\d+\.", env_vars["version"]) or env_vars["version"].startswith("/")):
+        if not (re.match(r"^\d+\.", config["version"]) or config["version"].startswith("/")):
             error("Version must either start with a number (e.g. '1.0.0') or be a an absolute path")
 
-        if not re.match(r"^[a-z0-9-]+$", env_vars["system_identifier"]):
+        if not re.match(r"^[a-z0-9-]+$", config["system_identifier"]):
             error("System identifier can only contain lowercase letters, numbers, and hyphens")
 
-        if not os.path.exists(env_vars["config_filepath"]):
-            error(f"Config source file not found: {env_vars['config_filepath']}")
+        if not os.path.exists(config["config_filepath"]):
+            error(f"Config source file not found: {config['config_filepath']}")
 
-        if not os.path.isfile(env_vars["config_filepath"]):
-            error(f"Config source must be a file, not a directory: {env_vars['config_filepath']}")
+        if not os.path.isfile(config["config_filepath"]):
+            error(f"Config source must be a file, not a directory: {config['config_filepath']}")
 
-        if env_vars["package_manager"] not in ["pip", "pdm", "uv"]:
+        if config["package_manager"] not in ["pip", "pdm", "uv"]:
             error("Package manager must be one of: pip, pdm, uv")
 
         if sys.version_info.major != 3 or sys.version_info.minor < 10:
             error("Python version must be 3.10 or higher")
 
-        print(json.dumps(env_vars, indent=4))
+        localprint(f"config = {json.dumps(config, indent=4)}")
 
-    return env_vars
-
-
-def setup_directories(config: dict[str, str]) -> tuple[str, str, str]:
-    with section("Setting up directories"):
[…]
