### 2025-09-04 d6dc71e
Improve setup scripts (18)
Refactor codebase ci test and reactivate all ci workflows
