### 2025-09-04 896ac97
Improve setup scripts (19)
Debug CI
Debug mosquitto scripts
