+++ scripts/mosquitto_stop.sh
@@ -1,7 +1,4 @@
 #!/bin/bash
 
-# fail on first error
-set -e errexit
-
 # stops the mosquitto broker
-pkill -f "mosquitto -d"
+pkill -f "mosquitto -d" || true
