+++ scripts/mosquitto_start.sh
@@ -10,7 +10,11 @@ export PROJECT_DIR=$(dirname $(dirname $(realpath $0)))
 mosquitto_passwd -c -b $PROJECT_DIR/tests/updater/data/mosquitto.passwords test_username test_password
 
 # configures the mosquitto broker
-echo "listener 1883\\nallow_anonymous false\\npassword_file $PROJECT_DIR/tests/updater/data/mosquitto.passwords" > $PROJECT_DIR/tests/updater/data/mosquitto.conf
+cat > $PROJECT_DIR/tests/updater/data/mosquitto.conf << EOF
+listener 1883
+allow_anonymous false
+password_file $PROJECT_DIR/tests/updater/data/mosquitto.passwords
+EOF
 
 # starts the mosquitto broker in the background
 mosquitto -d -c $PROJECT_DIR/tests/updater/data/mosquitto.conf
