### 2025-09-04 11a4971
Update setup scripts
Add option to configure development directory
Use this option in CI workflow as well
