### 2025-09-04 bcc8122
Debug setup scripts paths in CI
