+++ .github/workflows/test-production-setup.yaml
@@ -49,7 +49,7 @@ jobs:
           IVY_PACKAGE_MANAGER: pdm
           IVY_SYSTEM_IDENTIFIER: github-actions-node-development
           IVY_CONFIG_FILEPATH: ${{ github.workspace }}/config/config.template.json
-        run: curl -s https://raw.githubusercontent.com/${{ github.repository }}/refs/heads/${{ github.head_ref }}/scripts/setup_production.py | python3.10
+        run: curl -s https://raw.githubusercontent.com/${{ github.repository }}/${{ github.sha }}/scripts/setup_production.py | python3.10
 
       - name: Install mosquitto dependencies
         run: |
