### 2025-09-04 tum-esm/ivy#5 @dostuffthatmatters
Hi @ChitlangeSahas and @brpat07,

thank you again for the suggestions and issue reports!

**Setup Scripts:**

You can now run the following to set up a fully working dev environment. It will ask you for the path where you want to clone the repo to (doesn't have to be `~/Documents/{project_name}/dev` anymore):

```bash
curl -s https://raw.githubusercontent.com/tum-esm/ivy/refs/heads/main/scripts/setup_development.py | python3.12
```

There is a second script to fully set up a production node:

```bash
curl -s https://raw.githubusercontent.com/tum-esm/ivy/refs/heads/main/scripts/setup_production.py | python3.12
```

I updated the "Getting Started" guide accordingly. Both setup scripts are tested in GitHub actions now - including setting up the local MQTT broker and running the integration tests. In [`.github/workflows`](https://github.com/tum-esm/ivy/tree/main/.github/workflows) you can see the steps these tests are running through.

**Other Scripts:**

I added a bunch of scripts under [`scripts/`](https://github.com/tum-esm/ivy/tree/main/scripts) that perform various tasks when developing and testing the code. Before, they could be run with `pdm run ...`, but shell scripts are more established and don't require PDM to be installed.

Best,
Moritz
