### 2025-09-09 tum-esm/ivy#7 @brpat07
[Improvement]: Some improvements in scripts and documents
I tried to go through the repo as a new user and encountered few issues so maybe better to just improve them

1. You can have another .MD file which lists all the other docs or links for instructions and add that to [README](https://github.com/tum-esm/ivy/blob/main/README.md) -- Preferred
OR
Just add [Getting-started](https://tum-esm-ivy.netlify.app/getting-started) link to [README](https://github.com/tum-esm/ivy/blob/main/README.md) where user can easily find

2. The doc should say 'python' instead of 'python3.12'. (What if user has python 3.14 Or different version installed?)

3. Add a section for dependencies and add all the external dependencies (Eg, mosquitto, user can not run the software without it). I understand it is hard to run the script for every possible OS so just add the list which needs to be installed before starting with ivy. (Eg, python, mosquitto,
Why? Because when user runs the setup_development script and then run `mosquitto_start.sh` they may get this:
 ```
❯ ./scripts/mosquitto_start.sh
./scripts/mosquitto_start.sh: line 10: mosquitto_passwd: command not found
```
FYI, you can avoid adding instruction for these dependencies if the list is big; because otherwise you may have to add instructions for many OS. Eg, instead apt-get, since I am on mac I use brew. (But you can provide example if it would be useful)

4. add instruction to have 'IVY_ROOT_DIR' env variable. The setup_development script does not setup that env for all system. Eg, So if user open another tab or another terminal window to run `./scripts/run_tests_integration.sh` this would give user
```
raise FileNotFoundError(f"The directory {src.constants.ROOT_DIR} does not exist")
E   FileNotFoundError: The directory /Users/Tester/Documents/ivy does not exist
=========== short test summary info ============
ERROR tests/integration/test_connection_to_backend.py - FileNotFoundError: The directory /Users/bri...
!!!!!!!!!! stopping after 1 failures !!!!!!!!!!!
=============== 1 error in 0.19s ===============
```

5. Is there a dummy_sensor node I can test with? Once the setup is working, I started DAS with `python run.py`, now I want to check if DAS is able to receive data from the sensors? Apologize if this is already added but not able to easily find this so asking

(https://github.com/openjournals/joss-reviews/issues/8862)
