### 2025-09-09 tum-esm/ivy#5 @brpat07
@dostuffthatmatters few issues with script.

FYI, since I already had everything setup before, the script worked.
However, when I started as a new setup then I was getting some of the above issues. So you may want to try this as new setup.

#### Follow up in the setup_development script
1. Download and install command is failing
```
 ❯ curl -s https://raw.githubusercontent.com/tum-esm/ivy/refs/heads/main/scripts/setup_development.py | python3

What is the name of your DAS? (instead of `ivy`)
> Traceback (most recent call last):
  File "<stdin>", line 65, in <module>
  File "<stdin>", line 52, in get_validated_input
EOFError: EOF when reading a line
```
Solution: `curl -s https://raw.githubusercontent.com/tum-esm/ivy/refs/heads/main/scripts/setup_development.py > /tmp/setup_development.py && python /tmp/setup_development.py` maybe you can replace with this command

2. (Optional)
I would recommend to have default 'Where do you want to set up your dev repository? (if empty, uses `~/Documents/test_ivy/dev`)' as current directory, many people do not setup repos in ~/Documents (So you can get the pwd using os package in python)

3. (Optional)
It is asking for 'Where is your git repository hosted? (instead of `https://github.com/tum-esm/ivy`)'? Before this can we add an option to test the Ivy directly, if the user just wants to try/test IVY before adding to their repo?

4. Correct the example if just hyphen, lower-case and numbers are allowed
```
How do you want to name this computer? (if empty, uses "mac.attlocal.net")
> test.ivy.machine
Invalid input: Only lowercase letters, numbers, and hyphens are allowed
```

5. Script is failing at the end
```
❯ curl -s https://raw.githubusercontent.com/tum-esm/ivy/refs/heads/main/scripts/setup_development.py > /tmp/setup_development.py && python /tmp/setup_development.py
What is the name of your DAS? (instead of `ivy`)
> test_ivy
Where do you want to set up your dev repository? (if empty, uses `~/Documents/test_ivy/dev`)
> ~/Downloads/gito/test_ivy00
Where is your git repository hosted? (instead of `https://github.com/tum-esm/ivy`)
> https://github.com/tum-esm/ivy
How do you want to install your dependencies? (pip | pdm | uv)
> pip
How do you want to name this computer? (if empty, uses "mac.attlocal.net")
> test-ivy-machine
Which branch of the repository do you work on? (if empty, uses "main")
> main
Do you want to configure an MQTT broker? (yes | no)
> yes
Setting up your DAS with the following configuration:
  Project name:            test_ivy
  Git repository:          https://github.com/tum-esm/ivy
  Branch name:             main
  Package manager:         pip
  System identifier:       test-ivy-machine
  Configure MQTT broker:   yes

  Python Interpreter:      /Users/tester/.pyenv/versions/3.13.7/bin/python
  Installation Location:   ~/Downloads/gito/test_ivy00

Do you want to proceed? (yes | no)
> yes
Setting up directories: Starting
Setting up directories: Finished
...
... (Logs not needed so removed)
...
Installing dependencies using pip: Finished
Setting up configuration file: Starting
Traceback (most recent call last):
  File "/tmp/setup_development.py", line 215, in <module>
    with open(os.path.join(PROJECT_DIR, "config", "config.template.json"), "r") as f:
         ~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
FileNotFoundError: [Errno 2] No such file or directory: '~/Downloads/gito/test_ivy00/config/config.template.json'
```

Note:
- Points marked with '(Optional)' are recommendation not needed.
- Also, instead of just cluttering this issue, I created another issue https://github.com/tum-esm/ivy/issues/7 for some generla suggestion and failure I encounterd
- Otherwise it looks good.
