### 2025-09-23 7bd5dfe
Extend docs (2)
add contributor guide
