### 2025-09-23 f4e9715
Extend docs (3)
add code of conduct and contributors guide to root dir of repo
make sync script copy these files to docs
