### 2025-09-23 33ac69e
Improve docs (1)
Do not use a fixed Python version in the docs
