### 2025-09-23 be1f761
Write `.env` file during development setup, load `.env` during testing
