### 2025-09-23 f393744
Only allow updates in production environments or default paths
So that the updater doesn's accidentaly overwrite some user directories.
