### 2025-09-23 93d465f
Improve setup scripts (1)
Use current working directory as default location in setup script
