### 2025-09-23 ccca603
Improve setup scripts (2)
Allow the default location the dev setup uses in updater
