### 2025-09-23 fe641ce
Improve setup scripts (3)
Use the ivy repo as default dir in dev setup
