### 2025-09-23 fffa7c7
Improve setup scripts (5)
Make the default hostname in dev setup script comply with formatting rules
