### 2025-09-23 af1f5c4
Improve setup scripts (6)
Allow `~` in specified project dir
