### 2025-09-23 eb41f15
Improve setup scripts (7)
Possibly debug updater
