### 2025-09-23 526805f
Update contributors guide
