### 2025-09-24 tum-esm/ivy#5 @brpat07
@dostuffthatmatters great work. Looks like everything is working as expected and setup looks nice and organized. Thanks for updating. Closing this.
