### 2025-09-24 tum-esm/ivy#7 @brpat07
@dostuffthatmatters looks good, thanks for improvements.

> The "dummy procedure" and "system checks procedure" should already produce some dummy data without adding any code. Is that what you were looking for?

Sorry I missed it. (FYI, I tested this way. Ran DAS locally and check mqttt directly if the dummy_procedure was generating message and the messages are getting sent to mqttt by system, and I was able to verify system is sending walking-message (the test sensor in dummy_procedure) to message queue.)
