### 2025-09-25 8324f68
bib edits
Signed-off-by: Neea <neearusch@gmail.com>
