+++ paper/paper.bib
@@ -2,7 +2,7 @@
   title = {Networking the forest infrastructure towards near real-time monitoring – A white paper},
   volume = {872},
   ISSN = {0048-9697},
-  url = {\url{https://www.sciencedirect.com/science/article/abs/pii/S0048969723007830}},
+  url = {https://www.sciencedirect.com/science/article/abs/pii/S0048969723007830},
   DOI = {10.1016/j.scitotenv.2023.162167},
   journal = {Science of The Total Environment},
   publisher = {Elsevier BV},
@@ -14,10 +14,10 @@
 }
 
 @article{Bogena2022,
-  title = {COSMOS-Europe: a European network of cosmic-ray neutron soil moisture sensors},
+  title = {{COSMOS-Europe: a European network of cosmic-ray neutron soil moisture sensors}},
   volume = {14},
   ISSN = {1866-3516},
-  url = {\url{https://essd.copernicus.org/articles/14/1125/2022}},
+  url = {https://essd.copernicus.org/articles/14/1125/2022},
   DOI = {10.5194/essd-14-1125-2022},
   number = {3},
   journal = {Earth System Science Data},
@@ -29,10 +29,10 @@
 }
 
 @article{Dorigo2021,
-  title = {The International Soil Moisture Network: serving  Earth system science for over a decade},
+  title = {{The International Soil Moisture Network: serving Earth system science for over a decade}},
   volume = {25},
   ISSN = {1607-7938},
-  url = {\url{https://hess.copernicus.org/articles/25/5749/2021}},
+  url = {https://hess.copernicus.org/articles/25/5749/2021},
   DOI = {10.5194/hess-25-5749-2021},
   number = {11},
   journal = {Hydrology and Earth System Sciences},
@@ -44,10 +44,10 @@
 }
 
 @article{Zweifel2021,
-  title = {TreeNet–The Biological Drought and Growth Indicator Network},
+  title = {{TreeNet–The Biological Drought and Growth Indicator Network}},
   volume = {4},
   ISSN = {2624-893X},
-  url = {\url{https://www.frontiersin.org/articles/10.3389/ffgc.2021.776905/full}},
+  url = {https://www.frontiersin.org/articles/10.3389/ffgc.2021.776905/full},
   DOI = {10.3389/ffgc.2021.776905},
   journal = {Frontiers in Forests and Global Change},
   publisher = {Frontiers Media SA},
@@ -61,7 +61,7 @@
   title = {A Surface Ocean CO2 Reference Network,  SOCONET and Associated Marine Boundary Layer CO2 Measurements},
   volume = {6},
   ISSN = {2296-7745},
-  url = {\url{https://www.frontiersin.org/articles/10.3389/fmars.2019.00400/full}},
+  url = {https://www.frontiersin.org/articles/10.3389/fmars.2019.00400/full},
   DOI = {10.3389/fmars.2019.00400},
   journal = {Frontiers in Marine Science},
   publisher = {Frontiers Media SA},
@@ -71,10 +71,10 @@
 }
 
 @article{AlYaari2018,
-  title = {The AQUI Soil Moisture Network for Satellite Microwave Remote Sensing Validation in South-Western France},
+  title = {{The AQUI Soil Moisture Network for Satellite Microwave Remote Sensing Validation in South-Western France}},
   volume = {10},
   ISSN = {2072-4292},
-  url = {\url{https://www.mdpi.com/2072-4292/10/11/1839}},
+  url = {https://www.mdpi.com/2072-4292/10/11/1839},
   DOI = {10.3390/rs10111839},
   number = {11},
   journal = {Remote Sensing},
@@ -86,10 +86,10 @@
 }
 
 @article{Shusterman2016,
[…]
