### 2025-09-25 #8862 @nkrusch
:wave: @dostuffthatmatters,

I have reviewed the paper and offered some corrections in [PR #9](https://github.com/tum-esm/ivy/pull/9) (bib only). Please review the PR.

In addition, at this point I request that you go ahead and complete the following steps.

1. Make a (new) tagged release of the software.
1. Archive the reviewed software in Zenodo or Figshare, or a similar archival service.
1. Check the archival deposit has the correct metadata. This includes
   - title (should match the paper)
   - author list (should match the paper)
   - authors' ORCID (should match the paper)
   - release version tag should match the archive version name
      For example, for git tag `v1.0` the archive should be `v1.0`
     (on Zenodo the version string displays as "Version v1.0").

1. List here the version tag and DOI of the archive.

I can then move forward with recommending accept for the submission.
