+++ CITATION.cff
@@ -8,7 +8,7 @@ authors:
     given-names: Jia
     orcid: 0000-0002-6350-6610
 cff-version: 1.2.0
-date-released: "2024-12-29"
+date-released: "2025-09-25"
 doi: 10.5281/zenodo.14562882
 license:
   - cc-by-4.0
@@ -16,4 +16,4 @@ title:
   "Ivy: A Data Acquisition System for Distributed Sensor Networks Supporting
   Remote Configuration and Software Updates"
 type: software
-version: 1.0.0
+version: 1.1.0
