### 2025-09-25 6f78aa8
Fix API ref generation script
