### 2025-09-25 ee5e955
Update dev setup script
