### 2025-09-27 #8862 @dostuffthatmatters
Hi @danielskatz

v1 is the architecture, so all versions 1.X will follow the design described in this paper. v2 might change that. I can change it to v1.1 or remove v1 from the title if you prefer.
