### 2025-09-27 #8862 @danielskatz
I think removing v1 from the title would be cleanest, if you are ok doing this. I guess it also could stay as is, under the reasoning you describe
