### 2025-09-27 #8862 @dostuffthatmatters
@danielskatz

Yes, sure, I can remove the version from the title.

I believe "boilerplate" can also be used as a noun (https://en.wikipedia.org/wiki/Boilerplate_code). It is not only the code, but the architecture as well. I would prefer keeping it at "boilerplate".

I can remove the "owned" of course.
