### 2025-09-27 fe7e472
Update paper
Remove "v1" from title
Remove "owned"
