### 2025-10-01 17ff459
Update paper.bib
