+++ paper/paper.bib
@@ -44,7 +44,7 @@
 }
 
 @article{Zweifel2021,
-  title = {{TreeNet–The Biological Drought and Growth Indicator Network}},
+  title = {{TreeNet}–The Biological Drought and Growth Indicator Network},
   volume = {4},
   ISSN = {2624-893X},
   url = {https://www.frontiersin.org/articles/10.3389/ffgc.2021.776905/full},
@@ -71,7 +71,7 @@
 }
 
 @article{AlYaari2018,
-  title = {{The AQUI Soil Moisture Network for Satellite Microwave Remote Sensing Validation in South-Western France}},
+  title = {The AQUI Soil Moisture Network for Satellite Microwave Remote Sensing Validation in {S}outh-{W}estern {F}rance},
   volume = {10},
   ISSN = {2072-4292},
   url = {https://www.mdpi.com/2072-4292/10/11/1839},
@@ -86,7 +86,7 @@
 }
 
 @article{Shusterman2016,
-  title = {{The BErkeley Atmospheric CO$_\mathsf{2}$ Observation Network: initial evaluation}},
+  title = {The {B}Erkeley Atmospheric {CO}$_\mathsf{2}$ Observation Network: initial evaluation}},
   volume = {16},
   ISSN = {1680-7324},
   url = {https://acp.copernicus.org/articles/16/13449/2016},
@@ -135,12 +135,12 @@
     number = {84},
     pages = {5131},
     author = {Aigner, Patrick and Makowski, Moritz and Luther, Andreas and Dietrich, Florian and Chen, Jia},
-    title = {{Pyra: Automated EM27/SUN Greenhouse Gas Measurement Software}},
+    title = {Pyra: Automated {EM27/SUN} Greenhouse Gas Measurement Software},
     journal = {Journal of Open Source Software}
 }
 
 @article{Caubel2019,
-  title = {{A Distributed Network of 100 Black Carbon Sensors for 100 Days of Air Quality Monitoring in West Oakland, California}},
+  title = {A Distributed Network of 100 Black Carbon Sensors for 100 Days of Air Quality Monitoring in {W}est {O}akland, {C}alifornia},
   volume = {53},
   ISSN = {1520-5851},
   url = {http://dx.doi.org/10.1021/acs.est.9b00282},
@@ -211,7 +211,7 @@
 
 @software{hermes,
   author = {{TUM ESM} and Aigner, Patrick and Böhm, Felix and Frölich, Lars and Makowski, Moritz and Schmitt, Adrian and Chen, Jia},
-  title = {{Hermes, Professorship of Environmental Sensing and Modeling, Technical University of Munich}},
+  title = {Hermes},
   url = {https://github.com/tum-esm/hermes},
   version = {1.0.1},
   date = {2024-10-14}
@@ -219,7 +219,7 @@
 
 @incollection{IPCC_2021_WGI_SPM,
    title = {Summary for Policymakers},
-   booktitle = {{Climate Change 2021: The Physical Science Basis. Contribution of Working Group I to the Sixth Assessment Report of the Intergovernmental Panel on Climate Change}},
+   booktitle = {Climate Change 2021: The Physical Science Basis. Contribution of Working Group {I} to the Sixth Assessment Report of the Intergovernmental Panel on Climate Change},
    publisher = {Cambridge University Press},
    address = {Cambridge, United Kingdom and New York, NY, USA},
    url = {https://www.ipcc.ch/report/ar6/wg1/downloads/report/IPCC_AR6_WGI_SPM.pdf},
@@ -233,7 +233,7 @@
 
 @incollection{IPCC_2022_WGII_SPM,
[…]
