### 2025-10-01 #8862 @danielskatz
I've also suggested a bunch of changes in the bib file to make the cases align with JOSS's style, in https://github.com/tum-esm/ivy/pull/11
