### 2025-10-01 #8862 @dostuffthatmatters
@danielskatz Thank you for the updates to the bib file. I merged them and revised the usage of "boilerplate" in the manuscript in [the latest commit](https://github.com/tum-esm/ivy/commit/e9cb3e5afad6c85a5e09373bfb98967753ed863b).
