### 2025-10-02 #8862 @dostuffthatmatters
Hi @danielskatz - the proof looks good. Thank you.
