### 2025-10-03 #8862 @danielskatz
Congratulations to @dostuffthatmatters (Moritz Makowski) and co-author on your JOSS publication!!

JOSS completely depends on volunteers and we couldn't do this with you
