### 2025-10-03 #8862 @dostuffthatmatters
Dear all,

Thank you very much @danielskatz and @nkrusch for editing this!

A big thanks to @brpat07 @ChitlangeSahas and @cppraveen for reviewing this. Your comments improved the project a lot!

Best regards,
Moritz Makowski
